Add ceph pitr test
