+++ tests/spread/test_backups_pitr_ceph.py/task.yaml
@@ -1,0 +1,11 @@
+summary: test_backups_pitr_ceph.py
+environment:
+  TEST_MODULE: test_backups_pitr_ceph.py
+execute: |
+  tox run -e integration -- "tests/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
+artifacts:
+  - allure-results
+backends:
+  - -lxd-vm  # Requires CI secrets
+systems:
+  - -ubuntu-24.04-arm
